backend/requirements.txt
bcrypt==4.2.1
email_validator==2.2.0
email_validator==2.2.0
loguru==0.7.3
motor==3.7.0
pydantic==2.10.6
PyJWT==2.10.1
uvicorn==0.34.0